Make the `interface` parameter a required parameter.
We had a problem, when the interface name was not explicitly defined. The worker registered itself then under localhost (by default).
This caused however, that the scheduler could not ping the worker since the localhost address was not correct. (We have encountered this error on the cluster when scheduler and worker were not on the same machine.)